Move adapter out of the chatview.
